avidii
Avidii, an ed-tech platform founded in Switzerland, has expanded its footprint in India. The company formally launched in 2021, has now onboarded over 400 tutors across India and has achieved more than 5,000 downloads. It is currently active in four countries across the world, and aims to achieve one million users and downloads in India by the end of this year, according to an official release.  

Avidii offers expert teachers for immediate, on-demand and customised learning of Mathematics, Physics, Chemistry and Biology, along with Commerce and Arts subjects, primarily for students between the ages of 11 and 18. The learning app, with features for instant connect, caters to various educational boards including the Central Board of Secondary Education (CBSE), the Indian Certificate of Secondary Examination (ICSE), and State Boards, empowering students to achieve high scores in their examinations, the release mentioned.

Currently, Avidii operates two business models. The B2B segment involves partnerships with educational institutions and aims to enhance the learning experience within schools, colleges and training institutes, while the B2C segment enables direct interaction between students and tutors and creates an on-demand one-on-one deep learning experience for the students, it added. 

“Avidii has attracted over 400 tutors across India, with thousands of users already having tried the platform. We aim to achieve 1 million users by year-end and embark on our journey of growth within the country. We are looking to increase our tutor network extensively and are hopeful of achieving good traction soon,” Deepak Subbarao, co-founder, Avidii, said. “We have created a simple platform in Avidii wherein the tutors make sure that the answers and explanations are complete, precise and accurate. At Avidii, tutors use a variety of advanced teaching tools—such as digital whiteboard, graphic annotation features and sharing of learning materials—to create comprehensive learning sessions for the students,” he added.

Source link: https://www.financialexpress.com/education-2/avidii-an-ed-tech-platform-begins-operations-in-india-aims-to-onboard-tutors-extensively/3108987/

Website link: https://in.avidii.com/